chore(main/help-menu): replace help menu with experimental version
