+++ packages/main/src/plugin/help-menu/help-menu.ts
@@ -19,35 +19,17 @@
 import { inject, injectable } from 'inversify';
 
 import { IPCHandle } from '/@/plugin/api.js';
-import { type IConfigurationNode, IConfigurationRegistry } from '/@api/configuration/models.js';
 import { ActionKind, ItemInfo } from '/@api/help-menu.js';
 import product from '/@product.json' with { type: 'json' };
 
 @injectable()
 export class HelpMenu {
   constructor(
-    @inject(IConfigurationRegistry) private configurationRegistry: IConfigurationRegistry,
     @inject(IPCHandle)
     private readonly ipcHandle: IPCHandle,
   ) {}
 
   init(): void {
-    const helpMenuConfiguration: IConfigurationNode = {
-      id: 'preferences.experimental.helpMenu',
-      title: 'Experimental (Help menu)',
-      type: 'object',
-      properties: {
-        'helpMenu.useProductConfig': {
-          description: 'Replace help menu with the one defined in the product',
-          type: 'object',
-          default: import.meta.env.DEV ? {} : undefined,
-          experimental: {},
-        },
-      },
-    };
-
-    this.configurationRegistry.registerConfigurations([helpMenuConfiguration]);
-
     this.ipcHandle('help-menu:getItems', async (): Promise<ItemInfo[]> => {
       return this.getItems();
     });
