+++ packages/renderer/src/lib/help/HelpActions.svelte
@@ -1,16 +1,7 @@
 <script lang="ts">
-import HelpActionsExperimental from '/@/lib/help/HelpActionsExperimental.svelte';
-import { Items } from '/@/lib/help/HelpItems';
-
 import HelpActionsItems from './HelpActionsItems.svelte';
 
-const isExperimentalHelpEnabled = $derived(
-  await window.isExperimentalConfigurationEnabled('helpMenu.useProductConfig'),
-);
+const items = $derived(await window.helpMenuGetItems());
 </script>
 
-{#if isExperimentalHelpEnabled}
-  <HelpActionsExperimental />
-{:else}
-  <HelpActionsItems items={Items}/>
-{/if}
+<HelpActionsItems items={items}/>
