--- packages/renderer/src/lib/help/HelpActionsExperimental.svelte
@@ -1,7 +1,0 @@
-<script lang="ts">
-import HelpActionsItems from './HelpActionsItems.svelte';
-
-const items = $derived(await window.helpMenuGetItems());
-</script>
-
-<HelpActionsItems items={items}/>
